db/requirements.txt
Flask
Flask-SQLAlchemy
pandas
PyMySQL
cryptography
flask-cors